[Feat] checkRoomInfo 비즈니스 로직 구체화 #25

+++ SantaManito-iOS/SantaManito-iOS/Feature/MakeRoom/ViewModel/CheckRoomInfoViewModel.swift
@@ -9,55 +9,78 @@ import UIKit
 import Combine
 
 class CheckRoomInfoViewModel: ObservableObject {
-
+    
+    //MARK: - Action, State
+    
     enum Action {
-        case load
+        case onAppear
         case makeRoomButtonClicked
         case deleteMission(Mission)
         case copyInviteCode
     }
     
-    @Published var dueDateTime: Date = Date() //TODO: 서버통신으로 가져올 값 or dataBinding으로 가져올 값
-    @Published var remainingDays: Int = 5 //TODO: 서버통신으로 가져올 값 or dataBinding으로 가져올 값
-    @Published var inviteCode: String = "1A2B3C" //TODO: 서버통신으로 가져올 값 or dataBinding으로 가져올 값
-    
-    @Published private(set) var state = State(
-        isPresented: false,
-        dueDate: Date().toDueDateAndTime
-    )
-    
     struct State {
-        var isPresented: Bool
-        var dueDate: String
+        var isPresented: Bool = false
+        var dueDate: String = Date().toDueDateAndTime
     }
     
-    @Published var missionList: [Mission] = [
-        Mission(content: "손 잡기"),
-        Mission(content: "손 잡기2"),
-        Mission(content: "손 잡기3"),
-        Mission(content: "손 잡기4"),
-        Mission(content: "손 잡기5"),
-        Mission(content: "손 잡기6"),
-        Mission(content: "손 잡기7"),
-        Mission(content: "손 잡기8"),
-        Mission(content: "손 잡기"),
-        Mission(content: "손 잡기"),
-        Mission(content: "손 잡기")
-    ]
-
+    //MARK: - Dependency
+    
+    var roomService: EditRoomServiceType
+    @Published var roomInfo: MakeRoomInfo = MakeRoomInfo(
+        name: "",
+        remainingDays: 3,
+        dueDate: Date()
+    )
+    @Published var missionList: [Mission]
+    
+    //MARK: - Init
+    
+    init(roomInfo: MakeRoomInfo,
+         missionList: [Mission],
+         roomService: EditRoomServiceType
+    ) {
+        self.roomInfo = roomInfo
+        self.missionList = missionList
+        self.roomService = roomService
+    }
+    
+    //MARK: - Properties
+    
+    @Published private(set) var state = State()
+    private let cancelBag = CancelBag()
+    
+    @Published var inviteCode: String?
+    
+    //MARK: - Methods
+    
     func send(action: Action) {
+        weak var owner = self
+        guard let owner else { return }
+        
         switch action {
-        case .load:
-            configDuedata() // 서버통신 이후 날짜로 변환하는 코드 
-        case .makeRoomButtonClicked:
-            state.isPresented = true
+        case .onAppear:
+            updateDueDate() // 서버통신 이후 날짜로 변환하는 코드
+            
         case .deleteMission(let mission):
             if let index = missionList.firstIndex(where: { $0.id == mission.id }) {
                 missionList.remove(at: index)
             }
+            
+        case .makeRoomButtonClicked:
+            roomService.createRoom(roomInfo: roomInfo, missions: missionList)
+                .catch { _ in Empty() }
+                .sink { inviteCode in
+                    owner.inviteCode = inviteCode
+                    owner.updateDueDate()
+                    owner.state.isPresented = true
+                }
+                .store(in: cancelBag)
+            
         case .copyInviteCode:
             print("초대코드 복사")
             UIPasteboard.general.string = inviteCode
+            state.isPresented = false
             break
         }
     }
@@ -65,17 +88,11 @@ class CheckRoomInfoViewModel: ObservableObject {
 
 extension CheckRoomInfoViewModel {
     ///마니또 공개일을 계산하는 함수
-    func configDuedata() {
+    func updateDueDate() {
         // 날짜와 관련된 내용
-        let currentDate = Date()
-        guard let futureDate = Calendar.current.date(byAdding: .day, value: remainingDays, to: currentDate) else { return }
-        let dateComponents = Calendar.current.dateComponents([.year, .month, .day], from: futureDate)
-        guard let newDate = Calendar.current.date(from: dateComponents) else { return }
-        
-        
+        guard let futureDate = Calendar.current.date(byAdding: .day, value: roomInfo.remainingDays, to: Date()),
+              let newDate = Calendar.current.date(from: Calendar.current.dateComponents([.year, .month, .day], from: futureDate)) else { return }
         //시간과 관련된 내용
-        
-        
-        state.dueDate = newDate.toDueDate + " " + dueDateTime.toDueDateTime
+        state.dueDate = newDate.toDueDate + " " + roomInfo.dueDate.toDueDateTime
     }
 }
